development/tke-privacy-engine/functions/drive-watcher/requirements.txt
functions-framework>=3.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-cloud-pubsub>=2.0.0
google-cloud-storage>=2.0.0
cloudevents>=1.0.0
